requirements.txt
django-debug-toolbar
asgiref==3.7.2
Django==5.0.4
sqlparse==0.4.4
psycopg[binary]==3.1.12
Pillow
django-allauth
requests
pyotp
djangorestframework
python-dotenv
jwt
whitenoise
